refactor: move to shared types
