+++ domain/provisioner/internal/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2025 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// Package internal contains types to be shared within the provisioner
+// domain, but not outside of it.
+package internal
